Don't recommmend max_split_size_mb first
I've run into a couple of cases now where max_split_size_mb has been set
in projects as a workaround for fragmentation but it ends up causing problems
later, such as degraded performance from freeing empty segments. While it
is a useful setting to have, expandable_segments is probably a better first
resort for fixing fragmentation since when it works it is less likely to
need synchronous GPU operations to continue running.

ghstack-source-id: afd14d8b1441daed8017052388c1b9076a02951b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113481