[AOTI] Improve the two-pass wrapper codegen
Summary: For the second-pass, we don't have to rerun the whole inductor flow again. This PR moves that second-pass to the codegen time. This change not only speeds up the compilation, but also removes kernel scheduling inconsistency between the two passes. Another future improvement is to make the second-pass reuse the scheduler and do the wrapper codegen only.

ghstack-source-id: 40f58d193fb3424d162ad0be43ae69f2a8cbe691
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113762