Add mechanism for make_fx to not error on data-dependent-ops
I'm looking for a make_fx(tracing_mode=real) that doesn't error out on
data-dependent operations. This PR adds a flag to do that. We use this
to help implement offline generation, but this is useful by itself:
sometimes we want to trace a function with real tensors and don't care
if we bake values in (because we just want to see what happened).

Test Plan:
- new test

ghstack-source-id: 09593c96e741e2a40c7901102ffbbc05ca6aa5f3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114129