Add support for models with mutated buffer on torch.onnx.dynamo_export
This PR adds a unit test that leverages `torch.export.ExportedProgram`
models that mutates registered buffers. Although the exporter already
works out of the box in such scenario, the GraphModule and the exported
ONNX model have extra outputs containing the mutated buffers.
On future forward calls of the ONNX model, the mutated buffers are
used as input to the model.

The aforementioned extra inputs and outputs are by design and the
`ONNXProgram.model_signature` can be used to fetch detailed
input/output schema for the exported model.

However, when we want to compare pytorch output to ONNX's, there is
a mismatch between the schema because pytorch output does not include
the mutated buffers present on the ONNX output.

This PR extends
`onnx_program.adapt_torch_outputs_to_onnx(torch_outputs)` so that the
mutated buffers are prepended to the Pytorch output,
matching the ONNX schema.

ghstack-source-id: d8ea45be8673060828b2b6476673427333a2136f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112272